avoid ambiguous module bindings
The `compile-program` function generates syntax objects that
have bindings already. Adjust some symbol-to-identifier corecions
to ensure that the bindings are consistent, and adjust the context
where the output is used to avoid ambigious binding (in the case of
a macro system that preserves and layers contexts instead of ignoring
or stripping some).