Combine %rel clauses into a list
This fixes cuts in cases where %assert! and %assert-after!
are used. My interpretation of the documentation is that a
cut in a relation should skip _all_ following clauses, including
any introduced by assertions (and vice versa for assert-after).

As before, the relation itself is not mutated - only the
identifier is reassigned.

As a bonus, this makes relations significantly faster.