Fix %freeze/%melt{-new}
I am not sure if this is how they were originally envisioned to work, but I
found it surprising that they (at present) do not really support freezing terms
where there are free variables inside and you don't know where; I was only able
to get them to behave when each variable is always explicitly named in the
terms passed to %freeze and %melt, e.g.

    (%freeze (list X Y Z) (list *X *Y *Z))
    (%melt (list *X *Y *Z) (list A B C))

works just fine, but:

    (%freeze (list X Y Z) F)
    (%melt F (list A B C))

results in A, B, C getting bound to `<frozen>` structs, and not the original
variables. This is because the frozen struct gets folded into the overall term
F and will not be recognized as a `frozen-logic-var?` there.

---

The fix I'm proposing here is to represent frozen variables as `(make-frozen
<logic var>)` directly and then look for the `frozen?` structure in the
procedures that care about frozen terms. This has the special convenience that
in the rest of "unify.rkt", `frozen?` terms will now be recognized as `atom?`
which has the behavior we want anyways.

I added some tests to this effect; all but the first two would fail without the
corresponding changes to "unify.rkt".

A small style note: I opted not to add `frozen?` to `uni-match` since only the
triple (freeze,melt,melt-new) actually want to treat these differently from
`atom?`s